+++ common-libs/internal-api-clients/src/services/AffinidiVaultApiService.ts
@@ -5,7 +5,7 @@ import { ParseSpec } from '../types/openapiParser'
 import { BuildApiType } from '../types/typeBuilder'
 import GenericApiService from './GenericApiService'
 
-type ConstructorOptions = { vaultUrl: string; accessApiKey: string }
+type ConstructorOptions = { vaultUrl: string; accessApiKey: string, headers?: object }
 
 export type BlobType = {
   cyphertext: string