+++ sdk/core/src/_helpers.ts
@@ -1,6 +1,8 @@
 import { SUPPORTED_DID_METHODS } from './_defaultConfig'
 import SdkErrorFromCode from './shared/SdkErrorFromCode'
 
+import {version} from '../package.json';
+
 export const validateDidMethodSupported = (didMethod: string) => {
   if (!SUPPORTED_DID_METHODS.includes(didMethod)) {
     throw new SdkErrorFromCode('COM-10', { didMethod: didMethod, supportedDidMethods: SUPPORTED_DID_METHODS })
@@ -17,3 +19,5 @@ export const stripParamsFromDidUrl = (did: string): string =>
 export function isW3cCredential(credential: any): boolean {
   return !!credential.type
 }
+
+export const createApiServiceHeaders = () => ({ 'X-SDK-Version': version })