+++ sdk/core/src/services/AffinidiVaultStorageService.ts
@@ -4,7 +4,7 @@ import { DidAuthService } from '@affinidi/affinidi-did-auth-lib'
 
 import { IPlatformEncryptionTools } from '../shared/interfaces'
 import { VaultCredential } from '../dto/vault.dto'
-import { isW3cCredential } from '../_helpers'
+import { createApiServiceHeaders, isW3cCredential } from '../_helpers'
 
 type AffinidiVaultStorageOptions = {
   accessApiKey: string
@@ -30,7 +30,11 @@ export default class AffinidiVaultStorageService {
     this._didAuthService = didAuthService
     this._keysService = keysService
     this._platformEncryptionTools = platformEncryptionTools
-    this._vaultApiService = new AffinidiVaultApiService(options)
+    this._vaultApiService = new AffinidiVaultApiService({
+      vaultUrl: options.vaultUrl,
+      accessApiKey: options.accessApiKey,
+      headers: createApiServiceHeaders(),
+    })
   }
 
   private async _authorizeVcVault(): Promise<string> {